web_dashboard/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy>=2.0.41
psycopg[binary]>=3.2.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
python-dotenv==1.0.1
